requirements.txt
streamlit==1.31.0
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
pytz==2025.2
urllib3==2.5.0
watchdog==6.0.0
